fix hot reloading warnings
