+++ lib/components/SoundMaker/index.jsx
@@ -35,6 +35,7 @@ export class SoundMaker extends Component {
         },
         // reverb: {
         //   wet: 1,
+        //   impulse: '../../assets/audio/reverbs/Block Inside.wav'
         // },
         delay: {
           delayTime: .5,
