+++ lib/index.js
@@ -1,25 +1,28 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
+import { render } from 'react-dom'
 import { AppContainer as HotReloader } from 'react-hot-loader'
-import Root from './containers/Root'
+import Root from './components/Root'
 import './styles.scss'
+import configureStore from './store'
 
-ReactDOM.render(
+const { store, history } = configureStore({})
+
+render(
   <HotReloader>
-    <Root/>
+    <Root store={store} history={history} />
   </HotReloader>,
   document.getElementById('root')
-);
+)
 
 // Hot Module Replacement API
 if (module.hot) {
-  module.hot.accept('./containers/Root', () => {
-    const NextRoot = require('./containers/Root.jsx').default
-    ReactDOM.render(
+  module.hot.accept('./components/Root', () => {
+    const NextRoot = require('./components/Root/index.jsx').default
+    render(
       <HotReloader>
-        <NextRoot/>
+        <NextRoot store={store} history={history}/>
       </HotReloader>,
       document.getElementById('root')
-    );
-  });
+    )
+  })
 }
