+++ lib/reducers/index.js
@@ -2,11 +2,13 @@ import { combineReducers } from 'redux'
 import UserReducer from './UserReducer'
 import CompositionReducer from './CompositionReducer'
 import SoundReducer from './SoundReducer'
+import { routerReducer } from 'react-router-redux'
 
 const rootReducer = combineReducers({
+  router: routerReducer,
   UserReducer,
   CompositionReducer,
-  SoundReducer 
+  SoundReducer,
 })
 
 export default rootReducer
